#74: refactored inference types.ts
