+++ src/app/shared-services/inferences/engine/simple-engine.spec.ts
@@ -5,7 +5,7 @@ import { TrajectoryData } from 'src/app/model/trajectory'
 import { HomeInference, WorkInference } from './definitions'
 import { SimpleEngine } from './simple-engine'
 import * as fixtures from './simple-engine.spec.fixtures'
-import { IInferenceEngine, InferenceDefinition } from '../types'
+import { IInferenceEngine, InferenceDefinition } from './types'
 import haversine from 'haversine-distance'
 
 describe('inferences/SimpleEngine', () => {
